+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The changelog is available [on GitHub][2].
   Implement `hit status` command with pretty output.
 * [#67](https://github.com/kowainik/hit-on/issues/67):
   Implement `hit stash` and `hit unstash` commands.
+* [#77](https://github.com/kowainik/hit-on/issues/77):
+  Implement `hit clear` command.
 * [#81](https://github.com/kowainik/hit-on/issues/81):
   Implement `hit diff` command with pretty diff.
 * [#82](https://github.com/kowainik/hit-on/issues/82):